requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
crispy-bootstrap5==2025.6
django==4.2.29
django-crispy-forms==2.4
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-jazzmin==3.0.1
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-recurrence==1.11.1
django-scheduler==0.10.1
icalendar==6.3.2
jsonfield==3.1.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
six==1.17.0
sqlparse==0.5.5
swapper==1.4.0
typing-extensions==4.13.2
tzdata==2025.3
